+++ CHANGELOG.md
@@ -1,13 +1,18 @@
-## 0.0.0 (18 Ap 2021)
+## 0.0.3 (18 Apr 2021)
+
+-   Featured customizable theme
+-   Fixed styled-components dependencies as peerDependencies
+
+## 0.0.2 (18 Apr 2021)
 
 -   Fixed react and react-dom dependencies as peerDependencies
 -   Fixed main and type definition path
 
-## 0.0.1 (18 Ap 2021)
+## 0.0.1 (18 Apr 2021)
 
 -   Fixed Package post install script
 -   Fixed README.md demo image
 
-## 0.0.0 (18 Ap 2021)
+## 0.0.0 (18 Apr 2021)
 
 -   Initial public release